feat: close rpc connections when not in use - without close on dial requests (#6649)
Description
---
Added the ability to close RPC connections for a given peer:
- The RPC server can request sessions to be dropped.
- RPC session counts on the server are atomically managed with RPC
server sessions.
- When a client peer connection (final clone) drops, as a safety
precaution, all client RPC connections will be signalled to drop,
resulting in the server RPC sessions closing.
- Improved management of wallet connectivity to base nodes.
- **Edit:** Introduced an optional culling of the oldest client RPC
connection(s) into an RPC server for a given peer (_default: true_).
This will allow all new RPC connections to succeed. If older RPC
sessions are dropped by the server while the client is in use, the
client can re-request an RPC connection, but many times the clients are
not aware of the RPC sessions held on the server.

Motivation and Context
---
Wallets keep on hitting their maximum allowed RPC sessions in the
server.

How Has This Been Tested?
---
- Added new unit tests.
- System-level testing.

What process can a PR reviewer use to test or verify this change?
---
Code review

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>