fix(comms): ensure that inbound messaging terminates on disconnect (#6653)
Description
---
fix(comms): ensure that inbound messaging terminates on disconnect

Motivation and Context
---
Ensure that the inbound messaging worker terminates when the peer
connection is disconnected.
This PR ensures this by using the `PeerConnection::on_disconnect`
future. It has not been confirmed
that the inbound worker would not terminate before however this PR will
guarantee it.

How Has This Been Tested?
---
Updated unit tests. 

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->