feat: fix base node shutdown (#6697)
Description
---
Fixed the base node shutdown on Ubuntu (and hopefully on MacOS) whereas
pressing Ctrl-C multiple times rendered the command line interface in an
undefined state. This was not issue on Windows.

**Note:** The change to
'base_layer\core\src\base_node\tari_pulse_service\mod.rs' is an interim
change to enable testing (getting rid of the repeated command-line spam)
but not required for the base node to shut down. More changes in #6696.

Motivation and Context
---
The base node did not shut down properly on Ubuntu and MacOs.

How Has This Been Tested?
---
System-level testing on Windows and Ubuntu (WSL).

What process can a PR reviewer use to test or verify this change?
---
System-level testing - still outstanding on MacOS

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->