feat: new grpc method (#6742)
Description
---
Creates a new grpc method to get the reward of the nextblock, the sha +
rx estimate hash rates and metadata tip info
Adds caching to grpc calls

Motivation and Context
---
Currently, universe calls get_template to just read the reward which is
an expensive operation.
It also streams up to 100 headers to calculate the estimated hash rate. 

Both of these operations can be made much simpler and faster. This PR
provides a call to do just that.
This Pr also adds cahcing to the calls to only update if an update is
required.